+++ changelog.md
@@ -1,5 +1,18 @@
 ##Changelog
 
+**29/01/2016**
+- Chrolo Lib
+  - Added new functions to add tags to lines.
+  - Added "getLineInfo" which returns some simple line information like it's alignment, position on screen and frz. Takes into account style defaults and overrides
+- Line Splitter
+  - new macro to split at '\N', maintaining text position on screen
+  - Each new line attempts to maintain styling, including overrides present up to split point
+- Cycle Alignment
+  - Now detected current line alignment and cycles from there.
+  - Cycles each line's alignement individually.
+
+
+
 **13/01/2016**
 - Chrolo lib
   - getTextBound now correctly account for different fonts on the same line, outputting the correct size for the bounding box (Note: this fixes an issue with the Change Alignment macro)
@@ -27,6 +40,7 @@
  
 ##To do
 - [ ] Get fancy and put dep control into my stuff
+- [ ] Make the "Chrolo>¬macro¬" menu nesting optional
 
 ###Alignment Macro
 - [ ] have macro change \move tags appropriately aswell.
@@ -35,9 +49,10 @@
 
 ###Lib stuff
 - [ ] Improve chrolo.lib's TextBound function to handle: \shad, \fax, \fry, \frx, \shad... etc
-- [ ] Add ability to determine default line position when \pos not specified.
+- [x] Add ability to determine default line position when \pos not specified.
 - [ ] Turns out some tags, like \fad, will apply to the whole line no matter where they're placed. My current setup can't handle this...
-- [ ] make a function to add tag values to the first tag set of a line, and to add brackets if none yet exist.
+- [x] make a function to add tag values to the first tag set of a line, and to add brackets if none yet exist.
+- [x] Makes sure "subparts" return of linebound includes all overrides up to that point
 
 ###KFX lib stuff
 - [ ] Change how shuffle text's extra data is prepended